backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.26.0
python-docx==1.1.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
pywin32>=307
discord.py>=2.3.2
python-telegram-bot>=20.6
